Garnet

To obtain the list of implemented diffusion parameters for garnet, use:

GeoParams.MaterialParameters.ChemicalDiffusion.Garnet.chemical_diffusion_list — Function

chemical_diffusion_list(search::String="")

List all available chemical diffusion data for garnet. Includes an argument to search for a specific term, i.e. an element ("La") or an author.

The following diffusion parameters for garnet are implemented:

GeoParams.MaterialParameters.ChemicalDiffusion.Garnet.Grt_Mg_Chakraborty1992 — Function

Grt_Mg_Chakraborty1992()

Diffusion data of Mg in garnet. Calibrated between 1100-1480°C and 0.14-0.43 GPa in the C-O2 system. From Chakraborty and Ganguly (1992) (https://doi.org/10.1007/BF00296579) combined with data from Loomis et al. (1985) (https://doi.org/10.1007/BF00373040).

GeoParams.MaterialParameters.ChemicalDiffusion.Garnet.Grt_Mn_Chakraborty1992 — Function

Grt_Mn_Chakraborty1992()

Diffusion data of Mn in garnet. Calibrated between 1100-1480°C and 0.14-0.43 GPa in the C-O2 system. From Chakraborty and Ganguly (1992) (https://doi.org/10.1007/BF00296579) combined with data from Loomis et al. (1985) (https://doi.org/10.1007/BF00373040).

GeoParams.MaterialParameters.ChemicalDiffusion.Garnet.Grt_Fe_Chakraborty1992 — Function

Grt_Fe_Chakraborty1992()

Diffusion data of Fe in garnet. Calibrated between 1100-1480°C and 0.14-0.43 GPa in the C-O2 system. From Chakraborty and Ganguly (1992) (https://doi.org/10.1007/BF00296579) combined with data from Loomis et al. (1985) (https://doi.org/10.1007/BF00373040).

GeoParams.MaterialParameters.ChemicalDiffusion.Garnet.Grt_REE_Bloch2020_slow — Function

Grt_REE_Bloch2020_slow()

Diffusion data of REE in garnet for slow diffusion mechanism. Calibrated with experiments conducted between 950-1050°C at 1 atm with the QFM buffer and dry condition plus data from the literature. From Bloch et al. (2020) (https://doi.org/10.1093/petrology/egaa055) combined with data from Van Orman et al. (2002), Tirone et al. (2005) and Bloch et al. (2015).

GeoParams.MaterialParameters.ChemicalDiffusion.Garnet.Grt_REE_Bloch2020_fast — Function

Grt_REE_Bloch2020_fast()

Diffusion data of REE in garnet for fast diffusion mechanism. Calibrated with experiments conducted between 950-1050°C at 1 atm with the QFM buffer and dry condition plus data from the literature. From Bloch et al. (2020) (https://doi.org/10.1093/petrology/egaa055) combined with data from Van Orman et al. (2002), Tirone et al. (2005) and Bloch et al. (2015).