Rutile

To obtain the list of implemented diffusion parameters for rutile, use:

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.chemical_diffusion_list — Function

chemical_diffusion_list(search::String="")

List all available chemical diffusion data for rutile. Includes an argument to search for a specific term, i.e. an element ("La") or an author.

The following diffusion parameters for rutile are implemented:

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_3H_Caskey1974_perp_c — Function

Rt_3H_Caskey1974_perp_c

Diffusion data of Tritium in rutile. With anhydrous conditions and perpendicular to c-axis. Calibrated between 155-300°C. From Caskey et al. (1974) (https://doi.org/10.1016/0025-5416(74)90003-2).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_3H_Caskey1974_para_c — Function

Rt_3H_Caskey1974_para_c

Diffusion data of Tritium in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 155-300°C. From Caskey et al. (1974) (https://doi.org/10.1016/0025-5416(74)90003-2).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_3H_Cathcart1979_para_c — Function

Rt_3H_Cathcart1979_para_c

Diffusion data of Tritium in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 250-900°C. From Cathcart et al. (1979) (https://doi.org/110.1063/1.326490).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_3H_Cathcart1979_perp_c — Function

Rt_3H_Cathcart1979_perp_c

Diffusion data of Tritium in rutile. With anhydrous conditions and perpendicular to c-axis. Calibrated between 250-900°C. From Cathcart et al. (1979) (https://doi.org/110.1063/1.326490).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_H_Johnson1975_para_c — Function

Rt_H_Johnson1975_para_c

Diffusion data of H in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 125-750°C. From Johnson et al. (1975) (https://doi.org/10.1063/1.322206).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_H_Johnson1975_perp_c — Function

Rt_H_Johnson1975_perp_c

Diffusion data of H in rutile. With anhydrous conditions and perpendicular to c-axis. Calibrated between 125-750°C. From Johnson et al. (1975) (https://doi.org/10.1063/1.322206).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_He_Cherniak2011_para_c — Function

Rt_He_Cherniak2011_para_c

Diffusion data of He in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 250-500°C. From Cherniak et al. (2011) (https://doi.org/10.1016/j.chemgeo.2011.07.015).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_He_Cherniak2011_perp_c — Function

Rt_He_Cherniak2011_perp_c

Diffusion data of He in rutile. With anhydrous conditions and perpendicular to c-axis. Calibrated between 300-600°C. From Cherniak et al. (2011) (https://doi.org/10.1016/j.chemgeo.2011.07.015).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Li_Johnson1964_perp_c — Function

Rt_Li_Johnson1964_perp_c

Diffusion data of Li in rutile. With anhydrous conditions and perpendicular to c-axis. Calibrated between 80-360°C. From Johnson et al. (1964) (https://doi.org/10.1103/PhysRev.136.A284).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_O_Arita1979_para_c — Function

Rt_O_Arita1979_para_c

Diffusion data of O in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 1150-1450K. From Arita et al. (1979) (https://doi.org/10.1111/j.1151-2916.1979.tb19101.x).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_O_Arita1979_para_c_Cr — Function

Rt_O_Arita1979_para_c_Cr

Diffusion data of O in Cr-doped rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 1150-1450K. From Arita et al. (1979) (https://doi.org/10.1111/j.1151-2916.1979.tb19101.x).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_O_Dennis1993_perp_c — Function

Rt_O_Dennis1993_perp_c

Diffusion data of O in rutile. With hydrothermal conditions and perpendicular to c-axis. Calibrated between 873-1373K. From Dennis et al. (1993) (https://doi.org/10.1007/BF00414275).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_O_Dennis1993_perp_c_nat — Function

Rt_O_Dennis1993_perp_c_nat

Diffusion data of O in natural rutile. With hydrothermal conditions and perpendicular to c-axis. Calibrated between 873-1373K. From Dennis et al. (1993) (https://doi.org/10.1007/BF00414275).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_O_Derry1981_para_c — Function

Rt_O_Derry1981_para_c

Diffusion data of O in natural rutile. With parallel to c-axis. Calibrated between 1173-1673K. From Derry et al. (1981) (https://doi.org/10.1016/0022-3697(81)90011-1).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_O_Haul1965_unor — Function

Rt_O_Haul1965_unor

Diffusion data of O in rutile. Unoriented, anhydrous. Calibrated between 710-1300C. From Haul et al. (1965) (https://doi.org/10.1016/0022-3697(65)90066-1).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_O_Lundy1973_para_c — Function

Rt_O_Lundy1973_para_c

Diffusion data of O in natural rutile. With parallel to c-axis. Calibrated between 1200-1500K. From Lundy et al. (1973) (https://doi.org/10.1051/jphyscol:1973953).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_O_Lundy1973_perp_c — Function

Rt_O_Lundy1973_perp_c

Diffusion data of O in natural rutile. With perpendicular to c-axis. Calibrated between 1200-1500K. From Lundy et al. (1973) (https://doi.org/10.1051/jphyscol:1973953).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_O_Moore1998_para_c_fast — Function

Rt_O_Moore1998_para_c_fast

Diffusion data of O in natural rutile. With hydrous and anhydrous conditions, parallel to c-axis, at 0.1 to 1000 MPa. Calibrated between 1200-1500K. From Moore et al. (1998) (https://doi.org/10.2138/am-1998-7-803).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_O_Moore1998_para_c_slow — Function

Rt_O_Moore1998_para_c_slow

Diffusion data of O in natural rutile. With hydrous and anhydrous conditions, parallel to c-axis, at 0.1 to 1000 MPa. Calibrated between 1200-1500K. From Moore et al. (1998) (https://doi.org/10.2138/am-1998-7-803).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_O_Venkatu1970 — Function

Rt_O_Venkatu1970

Diffusion data of O in natural rutile. Calibrated between 900-1300C. From Venkatu et al. (1970).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Al_Cherniak2019_para_c — Function

Rt_Al_Cherniak2019_para_c

Diffusion data of Al in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 1100-1400°C. From Cherniak et al. (2019) (https://doi.org/10.2138/am-2019-7030).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Si_Cherniak2019_para_c — Function

Rt_Si_Cherniak2019_para_c

Diffusion data of Si in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 1100-1450°C. From Cherniak et al. (2019) (https://doi.org/10.2138/am-2019-7030).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Sc_Sasaki1985_para_c — Function

Rt_Sc_Sasaki1985_para_c

Diffusion data of Sc in rutile. With anhydrous conditions, in air and parallel to c-axis. Calibrated between 900-1500°C. From Sasaki et al. (1985) (https://doi.org/10.1016/0022-3697(85)90129-5).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Sc_Sasaki1985_perp_c — Function

Rt_Sc_Sasaki1985_perp_c

Diffusion data of Sc in rutile. With anhydrous conditions, in air and perpendicular to c-axis. Calibrated between 900-1100°C. From Sasaki et al. (1985) (https://doi.org/10.1016/0022-3697(85)90129-5).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Ti_Hoshino1985_para_c — Function

Rt_Ti_Hoshino1985_para_c

Diffusion data of Ti self-diffusion in rutile. Parallel c-axis, anhydrous. Calibrated between 1000-1500C.From Hoshino et al. (1985) (https://doi.org/10.1016/0022-3697(85)90079-4).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Ti_Hoshino1985_perp_c — Function

Rt_Ti_Hoshino1985_perp_c

Diffusion data of Ti self-diffusion in rutile. Perpendicular c-axis, anhydrous. Calibrated between 1000-1500C.From Hoshino et al. (1985) (https://doi.org/10.1016/0022-3697(85)90079-4).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Cr_Sasaki1985_para_c — Function

Rt_Cr_Sasaki1985_para_c

Diffusion data of Cr in rutile. With anhydrous conditions, in air and parallel to c-axis. Calibrated between 700-1400°C. From Sasaki et al. (1985) (https://doi.org/10.1016/0022-3697(85)90129-5).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Cr_Sasaki1985_perp_c — Function

Rt_Cr_Sasaki1985_perp_c

Diffusion data of Cr in rutile. With anhydrous conditions, in air and perpendicular to c-axis. Calibrated between 700-1400°C. From Sasaki et al. (1985) (https://doi.org/10.1016/0022-3697(85)90129-5).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Mn_Sasaki1985_para_c — Function

Rt_Mn_Sasaki1985_para_c

Diffusion data of Mn in rutile. With anhydrous conditions, in air and parallel to c-axis. Calibrated between 600-1000°C. From Sasaki et al. (1985) (https://doi.org/10.1016/0022-3697(85)90129-5).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Fe_Sasaki1985_para_c — Function

Rt_Fe_Sasaki1985_para_c

Diffusion data of Fe in rutile. With anhydrous conditions, in air and parallel to c-axis. Calibrated between 600-1400°C. From Sasaki et al. (1985) (https://doi.org/10.1016/0022-3697(85)90129-5).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Fe_Sasaki1985_perp_c — Function

Rt_Fe_Sasaki1985_perp_c

Diffusion data of Fe in rutile. With anhydrous conditions, in air and perpendicular to c-axis. Calibrated between 600-1400°C. From Sasaki et al. (1985) (https://doi.org/10.1016/0022-3697(85)90129-5).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Co_Sasaki1985_para_c — Function

Rt_Co_Sasaki1985_para_c

Diffusion data of Co in rutile. With anhydrous conditions, in air and parallel to c-axis. Calibrated between 500-900°C. From Sasaki et al. (1985) (https://doi.org/10.1016/0022-3697(85)90129-5).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Co_Sasaki1985_perp_c — Function

Rt_Co_Sasaki1985_perp_c

Diffusion data of Co in rutile. With anhydrous conditions, in air and perpendicular to c-axis. Calibrated between 800-1050°C. From Sasaki et al. (1985) (https://doi.org/10.1016/0022-3697(85)90129-5).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Ni_Sasaki1985_para_c — Function

Rt_Ni_Sasaki1985_para_c

Diffusion data of Ni in rutile. With anhydrous conditions, in air and parallel to c-axis. Calibrated between 700-900°C. From Sasaki et al. (1985) (https://doi.org/10.1016/0022-3697(85)90129-5).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Zr_Cherniak2007_para_c — Function

Rt_Zr_Cherniak2007_para_c

Diffusion data of Zr in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 750-1100°C. From Cherniak et al. (2007) (https://doi.org/10.1016/j.epsl.2007.06.027).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Nb_Marschall2013_para_c — Function

Rt_Nb_Marschall2013_para_c

Diffusion data of Nb in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 850-1250°C. From Marschall et al. (2013) (https://doi.org/10.1016/j.epsl.2013.05.055).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Hf_Cherniak2007_para_c — Function

Rt_Hf_Cherniak2007_para_c

Diffusion data of Hf in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 800-1000°C. From Cherniak et al. (2007) (https://doi.org/10.1016/j.epsl.2007.06.027).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Hf_Cherniak2007_perp_c — Function

Rt_Hf_Cherniak2007_perp_c

Diffusion data of Hf in rutile. With anhydrous conditions and perpendicular to c-axis. Calibrated between 750-1050°C. From Cherniak et al. (2007) (https://doi.org/10.1016/j.epsl.2007.06.027).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Ta_Marschall2013_para_c — Function

Rt_Ta_Marschall2013_para_c

Diffusion data of Ta in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 850-1250°C. From Cherniak et al. (2019) (https://doi.org/10.1016/j.epsl.2013.05.055).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Pb_Cherniak2000_unor — Function

Rt_Pb_Cherniak2000_unor

Diffusion data of Pb in rutile. With anhydrous conditions and unoriented. Calibrated between 700-1100°C. From Cherniak et al. (2000) (https://doi.org/10.1007/PL00007671).