Chemical Diffusion

Some routines and experimental data for chemical diffusion in minerals and phases are implemented in this module. Contributions are welcome to extend the database.

Methods

The following diffusion parameters are implemented:

Rutile:

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Zr_Cherniak2007_para_c — Function

Rt_Zr_Cherniak2007_para_c

Diffusion data of Zr in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 750-1100°C. From Cherniak et al. (2007) (https://doi.org/10.1016/j.epsl.2007.06.027).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Hf_Cherniak2007_para_c — Function

Rt_Hf_Cherniak2007_para_c

Diffusion data of Hf in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 800-1000°C. From Cherniak et al. (2007) (https://doi.org/10.1016/j.epsl.2007.06.027).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Hf_Cherniak2007_perp_c — Function

Rt_Hf_Cherniak2007_perp_c

Diffusion data of Hf in rutile. With anhydrous conditions and perpendicular to c-axis. Calibrated between 750-1050°C. From Cherniak et al. (2007) (https://doi.org/10.1016/j.epsl.2007.06.027).

Garnet:

GeoParams.MaterialParameters.ChemicalDiffusion.Garnet.Grt_Mg_Chakraborty1992 — Function

Grt_Mg_Chakraborty1992()

Diffusion data of Mg in garnet. Calibrated between 1100-1480°C and 0.14-0.43 GPa in the C-O2 system. From Chakraborty and Ganguly (1992) (https://doi.org/10.1007/BF00296579) combined with data from Loomis et al. (1985) (https://doi.org/10.1007/BF00373040).

GeoParams.MaterialParameters.ChemicalDiffusion.Garnet.Grt_Mn_Chakraborty1992 — Function

Grt_Mn_Chakraborty1992()

Diffusion data of Mn in garnet. Calibrated between 1100-1480°C and 0.14-0.43 GPa in the C-O2 system. From Chakraborty and Ganguly (1992) (https://doi.org/10.1007/BF00296579) combined with data from Loomis et al. (1985) (https://doi.org/10.1007/BF00373040).

GeoParams.MaterialParameters.ChemicalDiffusion.Garnet.Grt_Fe_Chakraborty1992 — Function

Grt_Fe_Chakraborty1992()

Diffusion data of Fe in garnet. Calibrated between 1100-1480°C and 0.14-0.43 GPa in the C-O2 system. From Chakraborty and Ganguly (1992) (https://doi.org/10.1007/BF00296579) combined with data from Loomis et al. (1985) (https://doi.org/10.1007/BF00373040).

GeoParams.MaterialParameters.ChemicalDiffusion.Garnet.Grt_REE_Bloch2020_slow — Function

Grt_REE_Bloch2020_slow()

Diffusion data of REE in garnet for slow diffusion mechanism. Calibrated with experiments conducted between 950-1050°C at 1 atm with the QFM buffer and dry condition plus data from the literature. From Bloch et al. (2020) (https://doi.org/10.1093/petrology/egaa055) combined with data from Van Orman et al. (2002), Tirone et al. (2005) and Bloch et al. (2015).

GeoParams.MaterialParameters.ChemicalDiffusion.Garnet.Grt_REE_Bloch2020_fast — Function

Grt_REE_Bloch2020_fast()

Diffusion data of REE in garnet for fast diffusion mechanism. Calibrated with experiments conducted between 950-1050°C at 1 atm with the QFM buffer and dry condition plus data from the literature. From Bloch et al. (2020) (https://doi.org/10.1093/petrology/egaa055) combined with data from Van Orman et al. (2002), Tirone et al. (2005) and Bloch et al. (2015).

Melt:

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Sc_Melt_Holycross2018_rhyolitic_highH2O — Function

Sc_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Sc in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Sc_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Sc_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Sc in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.V_Melt_Holycross2018_rhyolitic_highH2O — Function

V_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of V in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.V_Melt_Holycross2018_rhyolitic_mediumH2O — Function

V_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of V in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Y_Melt_Holycross2018_rhyolitic_highH2O — Function

Y_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Y in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Y_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Y_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Y in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Zr_Melt_Holycross2018_rhyolitic_highH2O — Function

Zr_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Zr in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Zr_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Zr_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Zr in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Hf_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Hf_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Hf in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Nb_Melt_Holycross2018_rhyolitic_highH2O — Function

Nb_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Nb in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Nb_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Nb_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Nb in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.La_Melt_Holycross2018_rhyolitic_highH2O — Function

La_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of La in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ni and Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.La_Melt_Holycross2018_rhyolitic_mediumH2O — Function

La_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of La in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni and Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Ce_Melt_Holycross2018_rhyolitic_highH2O — Function

Ce_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Ce in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Ce_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Ce_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Ce in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Pr_Melt_Holycross2018_rhyolitic_highH2O — Function

Pr_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Pr in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Pr_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Pr_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Pr in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Nd_Melt_Holycross2018_rhyolitic_highH2O — Function

Nd_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Nd in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Nd_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Nd_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Nd in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Sm_Melt_Holycross2018_rhyolitic_highH2O — Function

Sm_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Sm in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Sm_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Sm_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Sm in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Eu_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Eu_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Eu in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Gd_Melt_Holycross2018_rhyolitic_highH2O — Function

Gd_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Gd in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Gd_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Gd_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Gd in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Tb_Melt_Holycross2018_rhyolitic_highH2O — Function

Tb_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Tb in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Tb_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Tb_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Tb in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Dy_Melt_Holycross2018_rhyolitic_highH2O — Function

Dy_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Dy in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Dy_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Dy_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Dy in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Ho_Melt_Holycross2018_rhyolitic_highH2O — Function

Ho_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Ho in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Ho_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Ho_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Ho in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Er_Melt_Holycross2018_rhyolitic_highH2O — Function

Er_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Er in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Er_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Er_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Er in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Yb_Melt_Holycross2018_rhyolitic_highH2O — Function

Yb_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Yb in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Yb_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Yb_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Yb in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Lu_Melt_Holycross2018_rhyolitic_highH2O — Function

Lu_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Lu in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Lu_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Lu_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Lu in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Th_Melt_Holycross2018_rhyolitic_highH2O — Function

Th_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Th in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.U_Melt_Holycross2018_rhyolitic_highH2O — Function

U_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of U in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.U_Melt_Holycross2018_rhyolitic_mediumH2O — Function

U_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of U in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

Computational routines

To compute the diffusion coefficient from the parameters, use this:

GeoParams.MaterialParameters.ChemicalDiffusion.compute_D — Function

compute_D(data::DiffusionData; T=1K, P=0Pa, kwargs...)

Computes the diffusion coefficient D [m^2/s] from the diffusion data data at temperature T [K] and pressure P [Pa]. If T and P are provided without unit, the function assumes the units are in Kelvin and Pascal, respectively, and outputs the diffusion coefficient without unit based on the value in m^2/s.

GeoParams.MaterialParameters.ChemicalDiffusion.compute_D! — Function

compute_D!(D, data::DiffusionData; T=ones(size(D))K, P=zeros(size(D))Pa, kwargs...)

In-place version of compute_D(data::DiffusionData; T=1K, P=0Pa, kwargs...). D should be an array of the same size as T and P.

List of diffusion parameters

To know the available diffusion parameters for each phase, you can use X.chemical_diffusion_list(), where X is the phase of interest. For instance, for garnet, one can use Garnet.chemical_diffusion_list(). In addition, this function also takes an argument to search for a specific term, i.e. an element ("La") or an author.

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.chemical_diffusion_list — Function

chemical_diffusion_list(search::String="")

List all available chemical diffusion data for melt. Includes an argument to search for a specific term, i.e. an element ("La") or an author.

GeoParams.MaterialParameters.ChemicalDiffusion.Garnet.chemical_diffusion_list — Function

chemical_diffusion_list(search::String="")

List all available chemical diffusion data for melt. Includes an argument to search for a specific term, i.e. an element ("La") or an author.

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.chemical_diffusion_list — Function

chemical_diffusion_list(search::String="")

List all available chemical diffusion data for melt. Includes an argument to search for a specific term, i.e. an element ("La") or an author.