Melting Parameterizations

Methods

A number of melting parameterisations are implemented, which can be set with:

GeoParams.MeltingParam.MeltingParam_Caricchi — Type

MeltingParam_Caricchi()

Implements the T-dependent melting parameterisation used by Caricchi et al

\[ \theta = (800.0 .- (T + 273.15))./23.0 \]

\[ \phi_{solid} = 1.0 - {1.0 \over (1.0 + e^\theta)}; \]

Note that T is in Kelvin.

Computational routines

To compute the melt fraction at given T and P, use:

GeoParams.MeltingParam.compute_meltfraction! — Function

ComputeMeltingParam!(ϕ::AbstractArray{<:AbstractFloat}, P::AbstractArray{<:AbstractFloat},T:AbstractArray{<:AbstractFloat}, p::PhaseDiagram_LookupTable)

In-place computation of melt fraction in case we use a phase diagram lookup table. The table should have the collum :meltFrac specified.

GeoParams.MeltingParam.compute_meltfraction — Function

ComputeMeltingParam(P,T, p::AbstractPhaseDiagramsStruct)

Computes melt fraction in case we use a phase diagram lookup table. The table should have the collum :meltFrac specified.

ComputeMeltingParam!(ϕ::AbstractArray{<:AbstractFloat}, Phases::AbstractArray{<:Integer}, P::AbstractArray{<:AbstractFloat},T::AbstractArray{<:AbstractFloat}, MatParam::AbstractArray{<:AbstractMaterialParamsStruct})

In-place computation of density rho for the whole domain and all phases, in case a vector with phase properties MatParam is provided, along with P and T arrays.

Also note that phase diagrams can be imported using PerpleX_LaMEM_Diagram.