Plasticity

Plasticity is a non;linear rheology bthat is actrivated once stresses exceed a certain yield criteria.

Implemented laws

The following plastic law are implemented:

GeoParams.MaterialParameters.Plasticity.DruckerPrager — Type

DruckerPrager(ϕ=30, Ψ=0, C=10e6Pa)

Sets parameters for Drucker-Prager plasticity, where the yield stress $\sigma_{y}$ is computed by

\[ \sigma_{y} = (P-P_f)\tan(ϕ) + C\]

with $\phi$ being the friction angle (in degrees), $C$ cohesion, $P$ dynamic pressure and $P_f$ the fluid pressure (both positive under compression).

Yielding occurs when the second invariant of the deviatoric stress tensor, $\tau_{II}=(0.5\tau_{ij}\tau_{ij})^{0.5}$ touches the yield stress. This can be computed with the yield function $F$ and the plastic flow potential $Q$, which are respectively given by

\[ F = \tau_{II} - \cos(ϕ)C - \sin(ϕ)(P-P_f)\]

\[ Q = \tau_{II} - \sin(Ψ)(P-P_f) \]

Here, Ψ is the dilation angle, which must be zero for incompressible setups.

Plasticity is activated when $F(\tau_{II}^{trial})$ (the yield function computed with a trial stress) is >0. In that case, plastic strainrate $\dot{\varepsilon}^{pl}_{ij}$ is computed by:

\[ \dot{\varepsilon}^{pl}_{ij} =\dot{\lambda} {\partial Q \over \partial \sigma_{ij}}\]

where $\dot{\lambda}$ is a (scalar) that is nonzero and chosen such that the resuling stress gives $F(\tau_{II}^{final})=0$, and $\sigma_{ij}=-P + \tau_{ij}$ denotes the total stress tensor.

Computational routines

Once a plastic rheology is defined, we can use the following routines to perform computations within the solvers

GeoParams.MaterialParameters.Plasticity.compute_yieldfunction — Function

compute_yieldfunction(s::DruckerPrager; P, τII_old, Pf, kwargs...)

Computes the plastic yield function F for a given second invariant of the deviatoric stress tensor τII, P the pressure, and Pf fluid pressure.

GeoParams.MaterialParameters.Plasticity.compute_yieldfunction! — Function

compute_yieldfunction!(F::AbstractArray{_T,N}, s::DruckerPrager{_T}; P::AbstractArray{_T,N}, τII::AbstractArray{_T,N}, Pf=zero(P)::AbstractArray{_T,N}, kwargs...)

Computes the plastic yield function F for Drucker-Prager plasticity in an in-place manner. Required input arrays are pressure P and the second invariant of the deviatoric stress tensor τII at every point. You can optionally provide an array with fluid pressure Pf as well.