Seismic velocity

Methods

Seismic velocity can specified in a number of ways

GeoParams.MaterialParameters.SeismicVelocity.ConstantSeismicVelocity — Type

ConstantSeismicVelocity(Vp=8.1 km/s, Vs=4.5km/s)

Set a constant seismic P and S-wave velocity:

\[ V_p = cst\]

\[ V_s = cst\]

where $V_p, V_s$ are the P-wave and S-wave velocities [$km/s$].

In addition, you can use phase diagram lookup tables to compute seismic velocities as a function of pressure and temperature.

Computational routines

To evaluate seismic velocity within a user routine, use this:

Missing docstring.

Missing docstring for GeoParams.compute_swave_velocity!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GeoParams.compute_pwave_velocity!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GeoParams.compute_pwave_velocity. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GeoParams.compute_swave_velocity. Check Documenter's build log for details.

Seismic velocity correction for partial melt

Methods

The routine uses the reduction formulation of Clark et al., (2017) and is based on the equilibrium geometry model for the solid skeleton of Takei et al., 1998.

Computational routines

To compute melt-content based correction for seismic waves velocities, you can use:

Missing docstring.

Missing docstring for GeoParams.melt_correction. Check Documenter's build log for details.

Seismic S-wave velocity correction for (shallow depth) porosity

Methods

Ths routine is based on the equilibrium geometry model for the solid skeleton of Takei et al. (1998) and the porosity-depth empirical relationship of Chen et al. (2020)

Computational routines

To compute porosity based correction for seismic S-wave velocity, you can use:

Missing docstring.

Missing docstring for GeoParams.porosity_correction. Check Documenter's build log for details.

Seismic velocity correction for anelasticity

Methods

The routine uses the reduction formulation of karato (1993), using the quality factor formulation from Behn et al. (2009)

Computational routines

To compute a correction of S-wave velocity for anelasticity, use this:

GeoParams.MaterialParameters.SeismicVelocity.anelastic_correction — Function

    Vs_anel = anelastic_correction(water::Int64, Vs0::Float64,P::Float64,T::Float64)

This routine computes a correction of S-wave velocity for anelasticity

Input:

water: water flag, 0 = dry; 1 = dampened; 2 = water saturated

Vs0 : S-wave velocitiy of the solid phase (with or without melt correction)

P : pressure given in Pa

T : temperature given in °K

Output:

Vs_anel : corrected S-wave velocity for anelasticity

The routine uses the reduction formulation of Karato (1993), using the quality factor formulation from Behn et al. (2009)

References:

Karato, S. I. (1993). Importance of anelasticity in the interpretation of seismic tomography. Geophysical research letters, 20(15), 1623-1626.

Behn, M. D., Hirth, G., & Elsenbeck II, J. R. (2009). Implications of grain size evolution on the seismic structure of the oceanic upper mantle. Earth and Planetary Science Letters, 282(1-4), 178-189.