Zircon saturation parameterizations

Methods

Zircons are one of the ways in which we can date the age & activity of magmatic systems. Here, we implement simple parameterizations that estimates when zircons crystallize, which can be implemented in thermomechanical simulations.

GeoParams.MaterialParameters.ZirconSaturation.Tierney — Type

Tierney_etal()

The cummulative fraction of zircon crystallizing $F_{zr}$ is given by

\[ F_{zrs} = a - b \exp \left(-c \over T \right) \]

\[ F_{zrs} = 0 \textrm{ if } T<T_s\]

\[ F_{zrs} = 1 \textrm{ if }T>T_{zsat}\]

where $a$,$b$,$c$ are constants, $T$ the temperature in Kevin, $T_s$ is the solidus temperature, and $T_{zsat}$ the zircon saturation temperature.

The default parameters are from Tierney et al. (2016): $a=1.62$, $b=1.8*10^{-4}$, $c=1e4K$, $T_s=963.15 K$, $T_{zrs}=1098.15K$ Note that the supplement of Tierney et al., (2016) has a typo in the value of b, and that it is given in percentage, whereas here $0 \le F_{zrs} \le 1$.

Computational routines

To compute the zircon fraction Fzrs at given T and P, use:

Missing docstring.

Missing docstring for GeoParams.ZirconSaturation.compute_zirconsaturation!. Check Documenter's build log for details.

GeoParams.MaterialParameters.ZirconSaturation.compute_zirconsaturation — Function

compute_zirconsaturation!(Fzrs::AbstractArray{<:AbstractFloat}, Phases::AbstractArray{<:Integer}, T::AbstractArray{<:AbstractFloat}, MatParam::AbstractArray{<:AbstractMaterialParamsStruct})

In-place computation of zircon saturation Fzrs for the whole domain and all phases, in case a vector with phase properties MatParam is provided, along with T arrays.