Melt

To obtain the list of implemented diffusion parameters for melt, use:

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.chemical_diffusion_list — Function

chemical_diffusion_list(search::String="")

List all available chemical diffusion data for melt. Includes an argument to search for a specific term, i.e. an element ("La") or an author.

The following chemical diffusion parameters for melt are implemented:

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Sc_Melt_Holycross2018_rhyolitic_highH2O — Function

Sc_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Sc in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Sc_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Sc_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Sc in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.V_Melt_Holycross2018_rhyolitic_highH2O — Function

V_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of V in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.V_Melt_Holycross2018_rhyolitic_mediumH2O — Function

V_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of V in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Y_Melt_Holycross2018_rhyolitic_highH2O — Function

Y_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Y in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Y_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Y_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Y in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Zr_Melt_Holycross2018_rhyolitic_highH2O — Function

Zr_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Zr in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Zr_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Zr_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Zr in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Hf_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Hf_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Hf in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Nb_Melt_Holycross2018_rhyolitic_highH2O — Function

Nb_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Nb in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Nb_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Nb_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Nb in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.La_Melt_Holycross2018_rhyolitic_highH2O — Function

La_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of La in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ni and Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.La_Melt_Holycross2018_rhyolitic_mediumH2O — Function

La_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of La in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni and Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Ce_Melt_Holycross2018_rhyolitic_highH2O — Function

Ce_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Ce in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Ce_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Ce_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Ce in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Pr_Melt_Holycross2018_rhyolitic_highH2O — Function

Pr_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Pr in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Pr_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Pr_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Pr in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Nd_Melt_Holycross2018_rhyolitic_highH2O — Function

Nd_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Nd in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Nd_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Nd_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Nd in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Sm_Melt_Holycross2018_rhyolitic_highH2O — Function

Sm_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Sm in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Sm_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Sm_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Sm in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Eu_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Eu_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Eu in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Gd_Melt_Holycross2018_rhyolitic_highH2O — Function

Gd_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Gd in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Gd_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Gd_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Gd in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Tb_Melt_Holycross2018_rhyolitic_highH2O — Function

Tb_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Tb in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Tb_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Tb_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Tb in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Dy_Melt_Holycross2018_rhyolitic_highH2O — Function

Dy_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Dy in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Dy_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Dy_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Dy in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Ho_Melt_Holycross2018_rhyolitic_highH2O — Function

Ho_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Ho in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Ho_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Ho_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Ho in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Er_Melt_Holycross2018_rhyolitic_highH2O — Function

Er_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Er in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Er_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Er_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Er in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Yb_Melt_Holycross2018_rhyolitic_highH2O — Function

Yb_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Yb in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Yb_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Yb_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Yb in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Lu_Melt_Holycross2018_rhyolitic_highH2O — Function

Lu_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Lu in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Lu_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Lu_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Lu in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Th_Melt_Holycross2018_rhyolitic_highH2O — Function

Th_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Th in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.U_Melt_Holycross2018_rhyolitic_highH2O — Function

U_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of U in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.U_Melt_Holycross2018_rhyolitic_mediumH2O — Function

U_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of U in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).