Chemical Diffusion

Some routines and experimental data for chemical diffusion in minerals and phases are implemented in this module. Contributions are welcome to extend the database.

Computational routines

To compute the diffusion coefficient from the parameters, use this:

GeoParams.MaterialParameters.ChemicalDiffusion.compute_D — Function

compute_D(data::DiffusionData; T=1K, P=0Pa, fO2 = 1NoUnits, kwargs...)

Computes the diffusion coefficient D [m^2/s] from the diffusion data data at temperature T [K], pressure P [Pa] and oxygen fugacity fO2 [NoUnits]. If T and P are provided without unit, the function assumes the units are in Kelvin and Pascal, respectively, and outputs the diffusion coefficient without unit based on the value in m^2/s.

GeoParams.MaterialParameters.ChemicalDiffusion.compute_D! — Function

compute_D!(D, data::DiffusionData; T=ones(size(D))K, P=zeros(size(D))Pa, fO2 = ones(size(D)), kwargs...)

In-place version of compute_D(data::DiffusionData; T=1K, P=0Pa, fO2=0NoUnits, kwargs...). D should be an array of the same size as T, P and fO2.

Garnet

The following diffusion parameters for Garnet are implemented:

GeoParams.MaterialParameters.ChemicalDiffusion.Garnet.Grt_Mg_Chakraborty1992 — Function

Grt_Mg_Chakraborty1992()

Diffusion data of Mg in garnet. Calibrated between 1100-1480°C and 0.14-0.43 GPa in the C-O2 system. From Chakraborty and Ganguly (1992) (https://doi.org/10.1007/BF00296579) combined with data from Loomis et al. (1985) (https://doi.org/10.1007/BF00373040).

GeoParams.MaterialParameters.ChemicalDiffusion.Garnet.Grt_Mn_Chakraborty1992 — Function

Grt_Mn_Chakraborty1992()

Diffusion data of Mn in garnet. Calibrated between 1100-1480°C and 0.14-0.43 GPa in the C-O2 system. From Chakraborty and Ganguly (1992) (https://doi.org/10.1007/BF00296579) combined with data from Loomis et al. (1985) (https://doi.org/10.1007/BF00373040).

GeoParams.MaterialParameters.ChemicalDiffusion.Garnet.Grt_Fe_Chakraborty1992 — Function

Grt_Fe_Chakraborty1992()

Diffusion data of Fe in garnet. Calibrated between 1100-1480°C and 0.14-0.43 GPa in the C-O2 system. From Chakraborty and Ganguly (1992) (https://doi.org/10.1007/BF00296579) combined with data from Loomis et al. (1985) (https://doi.org/10.1007/BF00373040).

GeoParams.MaterialParameters.ChemicalDiffusion.Garnet.Grt_REE_Bloch2020_slow — Function

Grt_REE_Bloch2020_slow()

Diffusion data of REE in garnet for slow diffusion mechanism. Calibrated with experiments conducted between 950-1050°C at 1 atm with the QFM buffer and dry condition plus data from the literature. From Bloch et al. (2020) (https://doi.org/10.1093/petrology/egaa055) combined with data from Van Orman et al. (2002), Tirone et al. (2005) and Bloch et al. (2015).

GeoParams.MaterialParameters.ChemicalDiffusion.Garnet.Grt_REE_Bloch2020_fast — Function

Grt_REE_Bloch2020_fast()

Diffusion data of REE in garnet for fast diffusion mechanism. Calibrated with experiments conducted between 950-1050°C at 1 atm with the QFM buffer and dry condition plus data from the literature. From Bloch et al. (2020) (https://doi.org/10.1093/petrology/egaa055) combined with data from Van Orman et al. (2002), Tirone et al. (2005) and Bloch et al. (2015).

Melt

The following diffusion parameters for Melt are implemented:

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Sc_Melt_Holycross2018_rhyolitic_highH2O — Function

Sc_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Sc in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Sc_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Sc_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Sc in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.V_Melt_Holycross2018_rhyolitic_highH2O — Function

V_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of V in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.V_Melt_Holycross2018_rhyolitic_mediumH2O — Function

V_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of V in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Y_Melt_Holycross2018_rhyolitic_highH2O — Function

Y_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Y in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Y_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Y_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Y in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Zr_Melt_Holycross2018_rhyolitic_highH2O — Function

Zr_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Zr in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Zr_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Zr_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Zr in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Hf_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Hf_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Hf in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Nb_Melt_Holycross2018_rhyolitic_highH2O — Function

Nb_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Nb in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Nb_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Nb_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Nb in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.La_Melt_Holycross2018_rhyolitic_highH2O — Function

La_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of La in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ni and Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.La_Melt_Holycross2018_rhyolitic_mediumH2O — Function

La_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of La in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni and Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Ce_Melt_Holycross2018_rhyolitic_highH2O — Function

Ce_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Ce in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Ce_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Ce_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Ce in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Pr_Melt_Holycross2018_rhyolitic_highH2O — Function

Pr_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Pr in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Pr_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Pr_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Pr in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Nd_Melt_Holycross2018_rhyolitic_highH2O — Function

Nd_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Nd in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Nd_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Nd_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Nd in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Sm_Melt_Holycross2018_rhyolitic_highH2O — Function

Sm_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Sm in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Sm_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Sm_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Sm in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Eu_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Eu_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Eu in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Gd_Melt_Holycross2018_rhyolitic_highH2O — Function

Gd_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Gd in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Gd_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Gd_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Gd in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Tb_Melt_Holycross2018_rhyolitic_highH2O — Function

Tb_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Tb in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Tb_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Tb_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Tb in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Dy_Melt_Holycross2018_rhyolitic_highH2O — Function

Dy_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Dy in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Dy_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Dy_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Dy in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Ho_Melt_Holycross2018_rhyolitic_highH2O — Function

Ho_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Ho in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Ho_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Ho_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Ho in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Er_Melt_Holycross2018_rhyolitic_highH2O — Function

Er_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Er in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Er_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Er_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Er in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Yb_Melt_Holycross2018_rhyolitic_highH2O — Function

Yb_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Yb in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Yb_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Yb_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Yb in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Lu_Melt_Holycross2018_rhyolitic_highH2O — Function

Lu_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Lu in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Lu_Melt_Holycross2018_rhyolitic_mediumH2O — Function

Lu_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of Lu in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.Th_Melt_Holycross2018_rhyolitic_highH2O — Function

Th_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of Th in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.U_Melt_Holycross2018_rhyolitic_highH2O — Function

U_Melt_Holycross2018_rhyolitic_highH2O()

Diffusion data of U in rhyolitic melt (76.77 wt% SiO2) with 6.2 wt% of H2O. Calibrated with experiments conducted between 850-935°C at 1 GPa with Ag capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.U_Melt_Holycross2018_rhyolitic_mediumH2O — Function

U_Melt_Holycross2018_rhyolitic_mediumH2O()

Diffusion data of U in rhyolitic melt (76.77 wt% SiO2) with 4.1 wt% of H2O. Calibrated with experiments conducted between 960-1250°C at 1 GPa with Ni capsules from synthetic glass. From Holycross and Watson (2018) (https://doi.org/10.1016/j.gca.2018.04.006).

Rutile

The following diffusion parameters for Rutile are implemented:

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_3H_Caskey1974_perp_c — Function

Rt_3H_Caskey1974_perp_c

Diffusion data of Tritium in rutile. With anhydrous conditions and perpendicular to c-axis. Calibrated between 155-300°C. From Caskey et al. (1974) (https://doi.org/10.1016/0025-5416(74)90003-2).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_3H_Caskey1974_para_c — Function

Rt_3H_Caskey1974_para_c

Diffusion data of Tritium in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 155-300°C. From Caskey et al. (1974) (https://doi.org/10.1016/0025-5416(74)90003-2).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_3H_Cathcart1979_para_c — Function

Rt_3H_Cathcart1979_para_c

Diffusion data of Tritium in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 250-900°C. From Cathcart et al. (1979) (https://doi.org/110.1063/1.326490).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_3H_Cathcart1979_perp_c — Function

Rt_3H_Cathcart1979_perp_c

Diffusion data of Tritium in rutile. With anhydrous conditions and perpendicular to c-axis. Calibrated between 250-900°C. From Cathcart et al. (1979) (https://doi.org/110.1063/1.326490).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_H_Johnson1975_para_c — Function

Rt_H_Johnson1975_para_c

Diffusion data of H in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 125-750°C. From Johnson et al. (1975) (https://doi.org/10.1063/1.322206).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_H_Johnson1975_perp_c — Function

Rt_H_Johnson1975_perp_c

Diffusion data of H in rutile. With anhydrous conditions and perpendicular to c-axis. Calibrated between 125-750°C. From Johnson et al. (1975) (https://doi.org/10.1063/1.322206).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_He_Cherniak2011_para_c — Function

Rt_He_Cherniak2011_para_c

Diffusion data of He in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 250-500°C. From Cherniak et al. (2011) (https://doi.org/10.1016/j.chemgeo.2011.07.015).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_He_Cherniak2011_perp_c — Function

Rt_He_Cherniak2011_perp_c

Diffusion data of He in rutile. With anhydrous conditions and perpendicular to c-axis. Calibrated between 300-600°C. From Cherniak et al. (2011) (https://doi.org/10.1016/j.chemgeo.2011.07.015).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Li_Johnson1964_perp_c — Function

Rt_Li_Johnson1964_perp_c

Diffusion data of Li in rutile. With anhydrous conditions and perpendicular to c-axis. Calibrated between 80-360°C. From Johnson et al. (1964) (https://doi.org/10.1103/PhysRev.136.A284).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_O_Arita1979_para_c — Function

Rt_O_Arita1979_para_c

Diffusion data of O in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 1150-1450K. From Arita et al. (1979) (https://doi.org/10.1111/j.1151-2916.1979.tb19101.x).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_O_Arita1979_para_c_Cr — Function

Rt_O_Arita1979_para_c_Cr

Diffusion data of O in Cr-doped rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 1150-1450K. From Arita et al. (1979) (https://doi.org/10.1111/j.1151-2916.1979.tb19101.x).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_O_Dennis1993_perp_c — Function

Rt_O_Dennis1993_perp_c

Diffusion data of O in rutile. With hydrothermal conditions and perpendicular to c-axis. Calibrated between 873-1373K. From Dennis et al. (1993) (https://doi.org/10.1007/BF00414275).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_O_Dennis1993_perp_c_nat — Function

Rt_O_Dennis1993_perp_c_nat

Diffusion data of O in natural rutile. With hydrothermal conditions and perpendicular to c-axis. Calibrated between 873-1373K. From Dennis et al. (1993) (https://doi.org/10.1007/BF00414275).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_O_Derry1981_para_c — Function

Rt_O_Derry1981_para_c

Diffusion data of O in natural rutile. With parallel to c-axis. Calibrated between 1173-1673K. From Derry et al. (1981) (https://doi.org/10.1016/0022-3697(81)90011-1).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_O_Haul1965_unor — Function

Rt_O_Haul1965_unor

Diffusion data of O in rutile. Unoriented, anhydrous. Calibrated between 710-1300C. From Haul et al. (1965) (https://doi.org/10.1016/0022-3697(65)90066-1).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_O_Lundy1973_para_c — Function

Rt_O_Lundy1973_para_c

Diffusion data of O in natural rutile. With parallel to c-axis. Calibrated between 1200-1500K. From Lundy et al. (1973) (https://doi.org/10.1051/jphyscol:1973953).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_O_Lundy1973_perp_c — Function

Rt_O_Lundy1973_perp_c

Diffusion data of O in natural rutile. With perpendicular to c-axis. Calibrated between 1200-1500K. From Lundy et al. (1973) (https://doi.org/10.1051/jphyscol:1973953).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_O_Moore1998_para_c_fast — Function

Rt_O_Moore1998_para_c_fast

Diffusion data of O in natural rutile. With hydrous and anhydrous conditions, parallel to c-axis, at 0.1 to 1000 MPa. Calibrated between 1200-1500K. From Moore et al. (1998) (https://doi.org/10.2138/am-1998-7-803).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_O_Moore1998_para_c_slow — Function

Rt_O_Moore1998_para_c_slow

Diffusion data of O in natural rutile. With hydrous and anhydrous conditions, parallel to c-axis, at 0.1 to 1000 MPa. Calibrated between 1200-1500K. From Moore et al. (1998) (https://doi.org/10.2138/am-1998-7-803).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_O_Venkatu1970 — Function

Rt_O_Venkatu1970

Diffusion data of O in natural rutile. Calibrated between 900-1300C. From Venkatu et al. (1970).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Al_Cherniak2019_para_c — Function

Rt_Al_Cherniak2019_para_c

Diffusion data of Al in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 1100-1400°C. From Cherniak et al. (2019) (https://doi.org/10.2138/am-2019-7030).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Si_Cherniak2019_para_c — Function

Rt_Si_Cherniak2019_para_c

Diffusion data of Si in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 1100-1450°C. From Cherniak et al. (2019) (https://doi.org/10.2138/am-2019-7030).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Ti_Hoshino1985_para_c — Function

Rt_Ti_Hoshino1985_para_c

Diffusion data of Ti self-diffusion in rutile. Parallel c-axis, anhydrous. Calibrated between 1000-1500C.From Hoshino et al. (1985) (https://doi.org/10.1016/0022-3697(85)90079-4).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Ti_Hoshino1985_perp_c — Function

Rt_Ti_Hoshino1985_perp_c

Diffusion data of Ti self-diffusion in rutile. Perpendicular c-axis, anhydrous. Calibrated between 1000-1500C.From Hoshino et al. (1985) (https://doi.org/10.1016/0022-3697(85)90079-4).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Zr_Cherniak2007_para_c — Function

Rt_Zr_Cherniak2007_para_c

Diffusion data of Zr in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 750-1100°C. From Cherniak et al. (2007) (https://doi.org/10.1016/j.epsl.2007.06.027).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Nb_Marschall2013_para_c — Function

Rt_Nb_Marschall2013_para_c

Diffusion data of Nb in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 850-1250°C. From Cherniak et al. (2019) (https://doi.org/10.1016/j.epsl.2013.05.055).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Hf_Cherniak2007_para_c — Function

Rt_Hf_Cherniak2007_para_c

Diffusion data of Hf in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 800-1000°C. From Cherniak et al. (2007) (https://doi.org/10.1016/j.epsl.2007.06.027).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Hf_Cherniak2007_perp_c — Function

Rt_Hf_Cherniak2007_perp_c

Diffusion data of Hf in rutile. With anhydrous conditions and perpendicular to c-axis. Calibrated between 750-1050°C. From Cherniak et al. (2007) (https://doi.org/10.1016/j.epsl.2007.06.027).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Ta_Marschall2013_para_c — Function

Rt_Ta_Marschall2013_para_c

Diffusion data of Ta in rutile. With anhydrous conditions and parallel to c-axis. Calibrated between 850-1250°C. From Cherniak et al. (2019) (https://doi.org/10.1016/j.epsl.2013.05.055).

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.Rt_Pb_Cherniak2000_unor — Function

Rt_Pb_Cherniak2000_unor

Diffusion data of Pb in rutile. With anhydrous conditions and unoriented. Calibrated between 700-1100°C. From Cherniak et al. (2000) (https://doi.org/10.1007/PL00007671).

List of diffusion parameters

To know the available diffusion parameters for each phase, you can use X.chemical_diffusion_list(), where X is the phase of interest. For instance, for garnet, one can use Garnet.chemical_diffusion_list(). In addition, this function also takes an argument to search for a specific term, i.e. an element ("La") or an author.

GeoParams.MaterialParameters.ChemicalDiffusion.Rutile.chemical_diffusion_list — Function

chemical_diffusion_list(search::String="")

List all available chemical diffusion data for melt. Includes an argument to search for a specific term, i.e. an element ("La") or an author.

GeoParams.MaterialParameters.ChemicalDiffusion.Garnet.chemical_diffusion_list — Function

chemical_diffusion_list(search::String="")

List all available chemical diffusion data for melt. Includes an argument to search for a specific term, i.e. an element ("La") or an author.

GeoParams.MaterialParameters.ChemicalDiffusion.Melt.chemical_diffusion_list — Function

chemical_diffusion_list(search::String="")

List all available chemical diffusion data for melt. Includes an argument to search for a specific term, i.e. an element ("La") or an author.