Conductivity

Thermal conductivity is defined as

GeoParams.MaterialParameters.Conductivity.ConstantConductivity — Type

ConstantConductivity(k=3.0W/m/K)

Set a constant conductivity

\[ k = cst\]

where $k$ is the thermal conductivity [$W/m/K$].

GeoParams.MaterialParameters.Conductivity.T_Conductivity_Whittacker — Type

T_Conductivity_Whittacker()

Sets a temperature-dependent conductivity following the parameterization of Whittacker et al. (2009), Nature. Their parameterization is originally given for the thermal diffusivity, together with a parameterization for thermal conductivity, which allows us

\[ Cp = a + b T - c/T^2 \]

\[ \kappa = d/T - e, if T<=846K\]

\[ \kappa = f - g*T, if T>846K\]

\[ \rho = 2700 kg/m3\]

\[ k = \kappa Cp \rho\]

where $Cp$ is the heat capacity [$J/mol/K$], and $a,b,c$ are parameters that dependent on the temperature T:

a = 199.50 J/mol/K if T<= 846 K

a = 199.50 J/mol/K if T> 846 K

b = 0.0857J/mol/K^2 if T<= 846 K

b = 0.0323J/mol/K^2 if T> 846 K

c = 5e6J/mol*K if T<= 846 K

c = 47.9e-6J/mol*K if T> 846 K

d = 576.3m^2/s*K

e = 0.062m^2/s

f = 0.732m^2/s

g = 0.000135m^2/s/K

GeoParams.MaterialParameters.Conductivity.TP_Conductivity — Type

TP_Conductivity()

Sets a temperature (and pressure)-dependent conductivity parameterization as described in Gerya, Numerical Geodynamics (2nd edition, Table 21.2). The general for

\[ k = \left( a_k + {b_k \over {T + c_k}} \right) (1 + d_k P) \]

where $k$ is the conductivity [$W/K/m$], and $a_k,b_k,c_k,d_k$ are parameters that dependent on the temperature T and pressure P:

$a_k$ = 1.18Watt/K/m

$b_k$ = 474Watt/m

$c_k$ = 77K

$d_k$ = 0/MPa

GeoParams.MaterialParameters.Conductivity.Set_TP_Conductivity — Constant

Set_TP_Conductivity["Name of temperature(-pressure) dependent conductivity"]

This is a dictionary with pre-defined laws:

"UpperCrust"

"LowerCrust"

"OceanicCrust"

"Mantle"

Example

julia> k=Set_TP_Conductivity["Mantle"]
T/P dependent conductivity: k = (0.73 W K⁻¹ m⁻¹ + 1293 W m⁻¹/(T + 77 K))*(1 + 4.0e-5 MPa⁻¹*P)  

To compute, use this:

GeoParams.MaterialParameters.Conductivity.ComputeConductivity — Function

k = ComputeConductivity(P, T, s:<AbstractConductivity)

Returns the thermal conductivity k at any temperature T and pressure P using any of the parameterizations implemented.

Currently available:

ConstantConductivity

T_Conductivity_Whittacker

TP_Conductivity

Example

Using dimensional units

julia> T  = (250:100:1250)*K;
julia> cp = T_HeatCapacity_Whittacker()
julia> Cp = ComputeHeatCapacity(0,T,cp)

k = ComputeConductivity(T::Any, s::AbstractConductivity)

Computes conductivity if only temperature (and not pressure) is specified

k = ComputeConductivity(s::ConstantConductivity)

Returns conductivity if we are sure that we will only employ constant values throughout the simulation