CreepLaws

Implemented creep laws

The following viscous creep laws are implemented:

GeoParams.MaterialParameters.ConstitutiveRelationships.LinearViscous — Type

LinearViscous(η=1e20Pa*s)

Defines a linear viscous creeplaw

The (isotopic) linear viscous rheology is given by

\[ \tau_{ij} = 2 \eta \dot{\varepsilon}_{ij} \]

or

\[ \dot{\varepsilon}_{ij} = {\tau_{ij} \over 2 \eta }\]

where $\eta_0$ is the reference viscosity [Pa*s] at reference strain rate $\dot{\varepsilon}_0$[1/s], and $n$ the power law exponent [].

GeoParams.MaterialParameters.ConstitutiveRelationships.PowerlawViscous — Type

PowerlawViscous(η0=1e18Pa*s, n=2.0NoUnits, ε0=1e-15/s)

Defines a power law viscous creeplaw as:

\[ \tau_{ij}^n = 2 \eta_0 \left( \dot{\varepsilon}_{ij} \over \dot{\varepsilon}_0 \right)\]

where $\eta$ is the effective viscosity [Pa*s].

GeoParams.MaterialParameters.ConstitutiveRelationships.DislocationCreep — Type

DislocationCreep(n = 1.0NoUnits, r = 0.00.0NoUnits, A = 1.5MPa/s, E = 476.0kJ/mol, V = 6e-6m^3/mol, apparatus = AxialCompression )

Defines the flow law parameter of a dislocation creep law The (isotropic) dislocation creep law, as used by experimtalists, is given by

\[ \dot{\gamma} = A \sigma_\mathrm{d}^n f_\mathrm{H2O}^r \exp(-\frac{E+PV}{RT})\]

where $n$ is the power law exponent, $r$ is the exponent of fugacity dependence, $A$ is a pre-exponential factor MPa^(n+r), $E$ is the activation energy [kJ/mol], $V$ is the activation volume [m^3/mol]. $\dot{\gamma}$ is the ordinary strain rate [1/s], and $\sigma_\mathrm{d}$ is the differential stress which are converted into second invariants using the apparatus type that can be either AxialCompression, SimpleShear or Invariant. If the flow law paramters are already given as a function of second invariants, choose apparatus = "Invariant"

julia> x2      =   DislocationCreep(n=3)
DislocationCreep: n=3, r=0.0, A=1.5 MPa^-3 s^-1, E=476.0 kJ mol^-1, V=6.0e-6 m^3 mol^-1, Apparatus=AxialCompression

GeoParams.MaterialParameters.ConstitutiveRelationships.SetDislocationCreep — Function

SetDislocationCreep["Name of Dislocation Creep"]

Sets predefined dislocation creep data from a dictionary

Missing docstring.

Missing docstring for GeoParams.DiffusionCreep. Check Documenter's build log for details.

GeoParams.MaterialParameters.ConstitutiveRelationships.SetDiffusionCreep — Function

SetDiffusionCreep["Name of Diffusion Creep"]

This is a dictionary with pre-defined creep laws

Computational routines for creep laws

Once a creep rheology is defined, we can use the following routines to perform computations within the solvers

Missing docstring.

Missing docstring for CreepLawVariables. Check Documenter's build log for details.

GeoParams.MaterialParameters.ConstitutiveRelationships.compute_εII — Function

compute_εII(a::DiffusionCreep, TauII::_T; T::_T, P=one(_T), f=one(_T), d=one(_T), kwargs...)

Returns diffusion creep strainrate as a function of 2nd invariant of the stress tensor $\tau_{II}$

\[ \dot{ε}_{II} = A τ_{II}^n d^{p} f_{H_2O}^r \exp \left(- {{E + PV} \over RT} \right)\]

compute_εII(s::ConstantElasticity{_T}, τII; τII_old, dt)

Computes elastic strainrate given the deviatoric stress at the current (τII) and old timestep (τII_old), for a timestep dt:

\[ \dot{\varepsilon}^{el} = {1 \over 2 G} {D \tau_{II} \over Dt } ≈ {1 \over 2 G} {\tau_{II}- \tilde{\tau}_{II}^{old} \over dt }\]

Note that we here solve the scalar equation, which is sufficient for isotropic cases. In tensor form, it would be

\[ {\dot{\varepsilon}^{el}}_{ij} = {1 \over 2 G} { \tau_{ij} - \tilde{{\tau_{ij}}}^{old} \over dt }\]

here $\tilde{{\tau_{ij}}}^{old}$ is the rotated old deviatoric stress tensor to ensure objectivity (this can be done with Jaumann derivative, or also by using the full rotational formula).

τ = 2Gε + η̇ε =  2G ̇ε Δt + η̇ε
̇ε = τ/(2GΔt + 2η)

compute_εII(v::NTuple{N, AbstractConstitutiveLaw},τII,  args)

Compute deviatoric strain rate given deviatoric stress invariant

GeoParams.MaterialParameters.ConstitutiveRelationships.compute_εII! — Function

compute_εII!(EpsII::AbstractArray{_T,N}, a, TauII::AbstractArray{_T,N}; T, P, f,d,kwargs...)

Computes strainrate as a function of stress

compute_εII!(EpsII::AbstractArray{_T,N}, s::LinearViscous, TauII::AbstractArray{_T,N})

compute_εII!(ε_el::AbstractArray{_T,N}, s::ConstantElasticity{_T}; τII::AbstractArray{_T,N}, τII_old::AbstractArray{_T,N}, dt::_T, kwargs...)

In-place computation of the elastic shear strainrate for given deviatoric stress invariants at the previous (τII_old) and new (τII) timestep, as well as the timestep dt

\[ \dot{\varepsilon}^{el} = {1 \over 2 G} {D \tau_{II} \over Dt } ≈ {1 \over 2 G} {\tau_{II}- \tau_{II}^{old} \over dt }\]

GeoParams.MaterialParameters.ConstitutiveRelationships.compute_τII — Function

compute_τII(a::DislocationCreep, EpsII; P, T, f, args...)

Computes the stress for a Dislocation creep law given a certain strain rate

computeCreepLaw_TauII(EpsII::_T, a::DiffusionCreep; T::_T, P=zero(_T), f=one(_T), d=one(_T), kwargs...)

Returns diffusion creep stress as a function of 2nd invariant of the strain rate

compute_τII(s::LinearViscous, EpsII; kwargs...)

Returns second invariant of the stress tensor given a 2nd invariant of strain rate tensor

τ = 2Gε + η̇ε =  2G ̇ε Δt + η̇ε

compute_τII(v::NTuple{N, AbstractConstitutiveLaw}, εII, args; tol=1e-6, verbose=false)

Compute deviatoric stress invariant given strain rate 2nd invariant

GeoParams.MaterialParameters.ConstitutiveRelationships.compute_τII! — Function

compute_τII!(TauII::AbstractArray{_T,N}, a::DislocationCreep, EpsII::AbstractArray{_T,N}; 
    P =       zero(TauII)::AbstractArray{_T,N}, 
    T = ones(size(TauII))::AbstractArray{_T,N}, 
    f = ones(size(TauII))::AbstractArray{_T,N})

Computes the stress for a Dislocation creep law

compute_τII!(τII::AbstractArray{_T,N}, s::ConstantElasticity{_T}. ε_el::AbstractArray{_T,N}; τII_old::AbstractArray{_T,N}, dt::_T, kwargs...)

In-place update of the elastic stress for given deviatoric strainrate invariants and stres invariant at the old (τII_old) timestep, as well as the timestep dt

\[ \tau_{II} = 2 G dt \dot{\varepsilon}^{el} + \tau_{II}^{old}\]

Missing docstring.

Missing docstring for GeoParams.CorrectionFactor. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GeoParams.RemoveTensorCorrection. Check Documenter's build log for details.