CreepLaws

Implemented creep laws

The following viscous creep laws are implemented:

GeoParams.MaterialParameters.CreepLaw.LinearViscous — Type

LinearViscous(η=1e20Pa*s)

Defines a linear viscous creeplaw

The (isotopic) linear viscous rheology is given by

\[ \tau_{ij} = 2 \eta \dot{\varepsilon}_{ij} \]

or

\[ \dot{\varepsilon}_{ij} = {\tau_{ij} \over 2 \eta }\]

where $\eta_0$ is the reference viscosity [Pa*s] at reference strain rate $\dot{\varepsilon}_0$[1/s], and $n$ the power law exponent [].

GeoParams.MaterialParameters.CreepLaw.PowerlawViscous — Type

PowerlawViscous(η0=1e18Pa*s, n=2.0NoUnits, ε0=1e-15/s)

Defines a power law viscous creeplaw as:

\[ \tau_{ij}^n = 2 \eta_0 \left( \dot{\varepsilon}_{ij} \over \dot{\varepsilon}_0 \right)\]

where $\eta$ is the effective viscosity [Pa*s].

GeoParams.MaterialParameters.CreepLaw.DislocationCreep — Type

DislocationCreep(n = 1.0NoUnits, r = 0.00.0NoUnits, A = 1.5MPa/s, E = 476.0kJ/mol, V = 6e-6m^3/mol, apparatus = AxialCompression )

Defines the flow law parameter of a dislocation creep law The (isotropic) dislocation creep law, as used by experimtalists, is given by

\[ \dot{\gamma} = A \sigma_\mathrm{d}^n f_\mathrm{H2O}^r \exp(-\frac{E+PV}{RT})\]

where $n$ is the power law exponent, $r$ is the exponent of fugacity dependence, $A$ is a pre-exponential factor MPa^(n+r), $E$ is the activation energy [kJ/mol], $V$ is the activation volume [m^3/mol]. $\dot{\gamma}$ is the ordinary strain rate [1/s], and $\sigma_\mathrm{d}$ is the differential stress which are converted into second invariants using the apparatus type that can be either AxialCompression, SimpleShear or Invariant. If the flow law paramters are already given as a function of second invariants, choose apparatus = "Invariant"

julia> x2      =   DislocationCreep(n=3)
DislocationCreep: n=3, r=0.0, A=1.5 MPa^-3 s^-1, E=476.0 kJ mol^-1, V=6.0e-6 m^3 mol^-1, Apparatus=AxialCompression

GeoParams.MaterialParameters.CreepLaw.SetDislocationCreep — Function

SetDislocationCreep["Name of Dislocation Creep"]

This is a dictionary with pre-defined creep laws

Computational routines for creep laws

Once a creep rheology is defined, we can use the following routines to perform computations within the solvers

Missing docstring.

Missing docstring for CreepLawVariables. Check Documenter's build log for details.

GeoParams.MaterialParameters.CreepLaw.computeCreepLaw_EpsII — Function

computeCreepLaw_EpsII(TauII, s:<AbstractCreepLaw, p::CreepLawVariables)

Returns the strainrate invariant $\dot{\varepsilon}_{II}$ for a given deviatoric stress invariant $\tau_{II}$ for any of the viscous creep laws implemented. $p$ is a struct that can hold various parameters (such as $P$, $T$) that the creep law may need for the calculations

\[ \dot{\varepsilon}_{II} = f( \tau_{II} ) \]

GeoParams.MaterialParameters.CreepLaw.computeCreepLaw_TauII — Function

computeCreepLaw_TauII(EpsII, s:<AbstractCreepLaw, p::CreepLawVariables)

Returns the deviatoric stress invariant $\tau_{II}$ for a given strain rate invariant $\dot{\varepsilon}_{II}$ for any of the viscous creep laws implemented. $p$ is a struct that can hold various parameters (such as $P$, $T$) that the creep law may need for the calculations

\[ \tau_{II} = f( \dot{\varepsilon}_{II} ) \]