MaterialParameters

The material parameters structure holds all data for a given phase.

GeoParams.MaterialParameters.MaterialParams — Type

MaterialParams

Structure that holds all material parameters for a given phase

CreepLaws

Implemented creep laws

The following viscous creep laws are implemented:

Missing docstring.

Missing docstring for GeoParams.MaterialParameters.LinearViscous. Check Documenter's build log for details.

Once a creep rheology is defined, we can use the following routines to perform computations within the solvers

Missing docstring.

Missing docstring for GeoParams.MaterialParameters.CreepLaw_EpsII. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GeoParams.MaterialParameters.CreepLaw_TauII. Check Documenter's build log for details.

Computing

Density

The density equation of state can be specified in a number of ways