Melting Parameterizations

A number of melting parameterisations are implemented, which can be set with:

GeoParams.MeltingParam.MeltingParam_Caricchi — Type

MeltingParam_Caricchi()

Implements the T-dependent melting parameterisation used by Caricchi et al

\[ \theta = (800.0 .- (T + 273.15))./23.0 \]

\[ \phi_{solid} = 1.0 - {1.0 \over (1.0 + e^\theta)}; \]

Note that T is in Kelvin.

To compute the melt fraction at given T and P, use:

GeoParams.MeltingParam.ComputeMeltingParam — Function

ϕ  = ComputeMeltingParam(P,T, s:<MeltingParam_Caricchi)

Returns the melt fraction ϕ, for a given melting parameterization as a function of T,P

Also note that phase diagrams can be imported:

GeoParams.MaterialParameters.PhaseDiagrams.Read_LaMEM_Perple_X_Diagram — Function

PD_Data = Read_LaMEM_Perple_X_Diagram(fname::String; CharDim = nothing)

Reads a precomputed phase diagram in the LaMEM/Perple_X format (which is a phase diagram computed using Perple_X, but formatted in a manner that is readable using LaMEM). The data is stored in the PhaseDiagram_LookupTable structure.

If the CharDim object is specified, the values of all diagrams will be non-dimensionalized.

Example

===

julia> PD_Data = Read_LaMEM_Perple_X_Diagram("./test_data/Peridotite.in")
PerpleX_LaMEM Phase Diagram Lookup Table: 
   File    :   ./test_data/Peridotite.in
   T       :   293.0 - 1573.000039
   P       :   1.0e7 - 2.9999999944e9
   fields  :   meltRho
               meltFrac
               rockRho
               Rho
               rockVp
               rockVs
               rockVpVs

Once imported, the properties on the diagram can be interpolated in a simple manner:

julia> PD_Data.Rho(1500,1e7)
3042.836820256982

This also works for vectors or arrays:

julia> T = [1500 1800; 1233 1300]
julia> P = [1e8 1e9; 1e7 1e7]
julia> rho = PD_Data.Rho.(T,P)

(Note the dot . in front of the bracket while evaluating arrays).

The fields that are available depend on what is listed in the diagram file. The units of the fields are automatically evaluated, and employed to non-dimensionalize the parameters if CharDim is specified.

Algorithm

===

Internally, we employ linear interpolation, as provided by the Interpolations.jl package. Values outside the range of the diagram are set to the boundary of the diagram. The interpolation object is directly encoded in the PhaseDiagram_LookupTable` object.