Melting Parameterizations

Methods

A number of melting parameterisations are implemented, which can be set with:

GeoParams.MeltingParam.MeltingParam_Caricchi — Type

MeltingParam_Caricchi()

Implements the T-dependent melting parameterisation used by Caricchi et al

\[ \theta = {(800.0 - (T + 273.15)) \over 23.0} \]

\[ \phi_{melt} = {1.0 \over (1.0 + e^\theta)}\]

Note that T is in Kelvin.

GeoParams.MeltingParam.MeltingParam_5thOrder — Type

MeltingParam_5thOrder(a,b,c,d,e,f,T_s,T_l)

Uses a 5th order polynomial to describe the melt fraction phi between solidus temperature T_s and liquidus temperature T_l

\[ \phi = a T^5 + b T^4 + c T^3 + d T^2 + e T + f \textrm{ for } T_s ≤ T ≤ T_l\]

\[ \phi = 1 \textrm{ if } T>T_l\]

\[ \phi = 0 \textrm{ if } T<T_s\]

Temperature T is in Kelvin.

The default values are for a composite liquid-line-of-descent:

the upper part is for Andesite from: (Blatter, D. L. & Carmichael, I. S. (2001) Hydrous phase equilibria of a Mexican highsilica andesite: a candidate for a mantle origin? Geochim. Cosmochim. Acta 65, 4043–4065

the lower part is extrapolated to the granitic minimum using the Marxer & Ulmer LLD for Andesite (Marxer, F. & Ulmer, P. (2019) Crystallisation and zircon saturation of calc-alkaline tonalite from the Adamello Batholith at upper crustal conditions: an experimental study. Contributions Mineral. Petrol. 174, 84)

GeoParams.MeltingParam.MeltingParam_4thOrder — Type

MeltingParam_4thOrder(b,c,d,e,f,T_s,T_l)

Uses a 4th order polynomial to describe the melt fraction phi between solidus temperature T_s and liquidus temperature T_l

\[ \phi = b T^4 + c T^3 + d T^2 + e T + f \textrm{ for } T_s ≤ T ≤ T_l\]

\[ \phi = 1 \textrm{ if } T>T_l\]

\[ \phi = 0 \textrm{ if } T<T_s\]

Temperature T is in Kelvin.

The default values are for Tonalite experiments from Marxer and Ulmer (2019):

Marxer, F. & Ulmer, P. (2019) Crystallisation and zircon saturation of calc-alkaline tonalite from the Adamello Batholith at upper crustal conditions: an experimental study. Contributions Mineral. Petrol. 174, 84

GeoParams.MeltingParam.MeltingParam_Quadratic — Type

MeltingParam_Quadratic(T_s,T_l)

Quadratic melt fraction parameterisation where melt fraction $\phi$ depends only on solidus ($T_s$) and liquidus ($T_l$) temperature:

\[ \phi = 1.0 - \left( {T_l - T} \over {T_l - T_s} \right)^2\]

\[ \phi = 1.0 \textrm{ if } T>T_l \]

\[ \phi = 0.0 \textrm{ if } T<T_s \]

Temperature T is in Kelvin. This was used, among others, in Tierney et al. (2016) Geology

GeoParams.MeltingParam.MeltingParam_Assimilation — Type

MeltingParam_Assimilation(T_s,T_l,a)

Melt fraction parameterisation that takes the assimilation of crustal host rocks into account, as used by Tierney et al. (2016) based upon a parameterisation of Spera and Bohrson (2001)

Here, the fraction of molten and assimilated host rocks $\phi$ depends on the solidus ($T_s$) and liquidus ($T_l$) temperatures of the rocks, as well as on a parameter $a=0.005$

\[ X = \left( {T - T_s} \over {T_l - T_s} \right)\]

\[ \phi = a \cdot \left( \exp^{2ln(100)X} - 1.0 \right) \textrm{ if } X ≤ 0.5\]

\[ \phi = 1- a \cdot \exp^{2ln(100)(1-X)} \textrm{ if } X > 0.5\]

\[ \phi = 1.0 \textrm{ if } T>T_l \]

\[ \phi = 0.0 \textrm{ if } T<T_s \]

Temperature T is in Kelvin.

This was used, among others, in Tierney et al. (2016), who employed as default parameters:

\[ T_s=973.15, T_l=1173.15, a=0.005\]

References

Spera, F.J., and Bohrson, W.A., 2001, Energy-Constrained Open-System Magmatic Processes I: General Model and Energy-Constrained Assimilation and Fractional Crystallization (EC- AFC) Formulation: Journal of Petrology, v. 42, p. 999–1018.

Tierney, C.R., Schmitt, A.K., Lovera, O.M., de Silva, S.L., 2016. Voluminous plutonism during volcanic quiescence revealed by thermochemical modeling of zircon. Geology 44, 683–686. https://doi.org/10.1130/G37968.1

Computational routines

To compute the melt fraction at given T and P, use:

GeoParams.MeltingParam.compute_meltfraction! — Function

compute_meltfraction!(ϕ::AbstractArray{<:AbstractFloat}, P::AbstractArray{<:AbstractFloat},T:AbstractArray{<:AbstractFloat}, p::PhaseDiagram_LookupTable)

In-place computation of melt fraction in case we use a phase diagram lookup table. The table should have the column :meltFrac specified.

compute_meltfraction(ϕ::AbstractArray{<:AbstractFloat}, Phases::AbstractArray{<:Integer}, P::AbstractArray{<:AbstractFloat},T::AbstractArray{<:AbstractFloat}, MatParam::AbstractArray{<:AbstractMaterialParamsStruct})

In-place computation of melt fraction ϕ for the whole domain and all phases, in case an array with phase properties MatParam is provided, along with P and T arrays.

GeoParams.MeltingParam.compute_meltfraction — Function

compute_meltfraction(P,T, p::AbstractPhaseDiagramsStruct)

Computes melt fraction in case we use a phase diagram lookup table. The table should have the column :meltFrac specified.

ϕ = compute_meltfraction(Phases::AbstractArray{<:Integer}, P::AbstractArray{<:AbstractFloat},T::AbstractArray{<:AbstractFloat}, MatParam::AbstractArray{<:AbstractMaterialParamsStruct})

Computation of melt fraction ϕ for the whole domain and all phases, in case an array with phase properties MatParam is provided, along with P and T arrays.

You can also obtain the derivative of melt fraction versus temperature with (useful to compute latent heat effects):

GeoParams.MeltingParam.compute_dϕdT! — Function

compute_dϕdT!(dϕdT::AbstractArray{<:AbstractFloat}, P::AbstractArray{<:AbstractFloat},T:AbstractArray{<:AbstractFloat}, p::PhaseDiagram_LookupTable)

In-place computation of melt fraction in case we use a phase diagram lookup table. The table should have the column :meltFrac specified. The derivative is computed by finite differencing.

compute_dϕdT!(ϕ::AbstractArray{<:AbstractFloat}, Phases::AbstractArray{<:Integer}, P::AbstractArray{<:AbstractFloat},T::AbstractArray{<:AbstractFloat}, MatParam::AbstractArray{<:AbstractMaterialParamsStruct})

Computates the derivative of melt fraction ϕ versus temperature T, $\partial \phi} \over {\partial T}$ for the whole domain and all phases, in case an array with phase properties MatParam is provided, along with P and T arrays. This is employed in computing latent heat terms in an implicit manner, for example

GeoParams.MeltingParam.compute_dϕdT — Function

compute_dϕdT(P,T, p::AbstractPhaseDiagramsStruct)

Computes derivative of melt fraction vs T in case we use a phase diagram lookup table. The table should have the column :meltFrac specified. The derivative is computed by finite differencing.

ϕ = compute_dϕdT(Phases::AbstractArray{<:Integer}, P::AbstractArray{<:AbstractFloat},T::AbstractArray{<:AbstractFloat}, MatParam::AbstractArray{<:AbstractMaterialParamsStruct})

Computates the derivative of melt fraction ϕ versus temperature T for the whole domain and all phases, in case an array with phase properties MatParam is provided, along with P and T arrays. This is employed in computing latent heat terms in an implicit manner, for example

Also note that phase diagrams can be imported using PerpleX_LaMEM_Diagram, which may also have melt content information. The computational routines work with that as well.

Plotting routines

You can use the routine PlotMeltFraction to create a plot, provided that the Plots package has been loaded

Missing docstring.

Missing docstring for GeoParams.Plotting.PlotMeltFraction. Check Documenter's build log for details.