Melting Parameterizations

Methods

A number of melting parameterisations are implemented, which can be set with:

GeoParams.MeltingParam.MeltingParam_Caricchi — Type

MeltingParam_Caricchi()

Implements the T-dependent melting parameterisation used by Caricchi et al

\[ \theta = (800.0 .- (T + 273.15))./23.0 \]

\[ \phi_{solid} = 1.0 - {1.0 \over (1.0 + e^\theta)}; \]

Note that T is in Kelvin.

GeoParams.MeltingParam.MeltingParam_5thOrder — Type

MeltingParam_5thOrder(a,b,c,d,e,f,T_s,T_l)

Uses a 5th order polynomial to describe the melt fraction phi between solidus temperature T_s and liquidus temperature T_l

\[ \phi = a T^5 + b T^4 + c T^3 + d T^2 + e T + f \textrm{ if } T_s ≤ T ≤ T_l\]

\[ \phi = 1 \textrm{ if } T>T_l\]

\[ \phi = 0 \textrm{ if } T<T_s\]

Temperature T is in Kelvin.

The default values are for a composite liquid-line-of-descent:

the upper part is for Andesite from: (Blatter, D. L. & Carmichael, I. S. (2001) Hydrous phase equilibria of a Mexican highsilica andesite: a candidate for a mantle origin? Geochim. Cosmochim. Acta 65, 4043–4065

the lower part is extrapolated to the granitic minimum using the Marxer & Ulmer LLD for Andesite (Marxer, F. & Ulmer, P. (2019) Crystallisation and zircon saturation of calc-alkaline tonalite from the Adamello Batholith at upper crustal conditions: an experimental study. Contributions Mineral. Petrol. 174, 84)

GeoParams.MeltingParam.MeltingParam_4thOrder — Type

MeltingParam_4thOrder(b,c,d,e,f,T_s,T_l)

Uses a 4th order polynomial to describe the melt fraction phi between solidus temperature T_s and liquidus temperature T_l

\[ \phi = b T^4 + c T^3 + d T^2 + e T + f \textrm{ if } T_s ≤ T ≤ T_l\]

\[ \phi = 1 \textrm{ if } T>T_l\]

\[ \phi = 0 \textrm{ if } T<T_s\]

Temperature T is in Kelvin.

The default values are for Tonalite experiments from Marxer and Ulmer (2019):

Marxer, F. & Ulmer, P. (2019) Crystallisation and zircon saturation of calc-alkaline tonalite from the Adamello Batholith at upper crustal conditions: an experimental study. Contributions Mineral. Petrol. 174, 84

GeoParams.MeltingParam.MeltingParam_Quadratic — Type

MeltingParam_Quadratic(T_s,T_l)

Quadratic melt fraction parameterisation where melt fraction $\phi$ depends only on solidus ($T_s$) and liquidus ($T_l$) temperature:

\[ \phi = 1.0 - \left( {T_l - T} \over {T_l - T_s} \right)^2\]

\[ \phi = 1.0 \textrm{ if } T>T_l \]

\[ \phi = 0.0 \textrm{ if } T<T_s \]

Temperature T is in Kelvin. This was used, among others, in Tierney et al. (2016) Geology

Computational routines

To compute the melt fraction at given T and P, use:

GeoParams.MeltingParam.compute_meltfraction! — Function

ComputeMeltingParam!(ϕ::AbstractArray{<:AbstractFloat}, P::AbstractArray{<:AbstractFloat},T:AbstractArray{<:AbstractFloat}, p::PhaseDiagram_LookupTable)

In-place computation of melt fraction in case we use a phase diagram lookup table. The table should have the collum :meltFrac specified.

GeoParams.MeltingParam.compute_meltfraction — Function

ComputeMeltingParam(P,T, p::AbstractPhaseDiagramsStruct)

Computes melt fraction in case we use a phase diagram lookup table. The table should have the collum :meltFrac specified.

ComputeMeltingParam!(ϕ::AbstractArray{<:AbstractFloat}, Phases::AbstractArray{<:Integer}, P::AbstractArray{<:AbstractFloat},T::AbstractArray{<:AbstractFloat}, MatParam::AbstractArray{<:AbstractMaterialParamsStruct})

In-place computation of density rho for the whole domain and all phases, in case a vector with phase properties MatParam is provided, along with P and T arrays.

Also note that phase diagrams can be imported using PerpleX_LaMEM_Diagram.

Plotting routines

You can use the routine PlotMeltFraction to create a plot, provided that the Plots package has been loaded

Missing docstring.

Missing docstring for GeoParams.Plotting.PlotMeltFraction. Check Documenter's build log for details.