Melting Parameterizations

A number of melting parameterisations are implemented, which can be set with:

GeoParams.MeltingParam.MeltingParam_Caricchi — Type

MeltingParam_Caricchi()

Implements the T-dependent melting parameterisation used by Caricchi et al

\[ \theta = (800.0 .- (T + 273.15))./23.0 \]

\[ \phi_{solid} = 1.0 - {1.0 \over (1.0 + e^\theta)}; \]

Note that T is in Kelvin.

To compute the melt fraction at given T and P, use:

GeoParams.MeltingParam.ComputeMeltingParam — Function

ComputeMeltingParam(P,T, p::AbstractPhaseDiagramsStruct)

Computes melt fraction in case we use a phase diagram lookup table. The table should have the collum :meltFrac specified.

ϕ  = ComputeMeltingParam(P,T, s:<MeltingParam_Caricchi)

Returns the melt fraction ϕ, for a given melting parameterization as a function of T,P

Also note that phase diagrams can be imported using Read_LaMEM_Perple_X_Diagram.