Phase Diagrams

It is possible to employ phase diagrams as lookup tables, which can be used to compute density, melt fraction or seismic properties, for example.

Perple_X

A popular way to compute phase diagrams (pseudosections) for a given (fixed) chemical composition as a function of pressure and temperature is by using Perple_X. Once a stable assemblage is computed, through Gibbs energy minimisation, it allows you to output a large number of properties such as density, melt fraction or seismic velocities.

Within GeoParams, we can import this diagram (provided it is formatted in the same manner as LaMEM expects Perple_X input to be).

Missing docstring.

Missing docstring for GeoParams.MaterialParameters.PhaseDiagrams.Read_LaMEM_Perple_X_Diagram. Check Documenter's build log for details.

GeoParams.MaterialParameters.PhaseDiagrams.PhaseDiagram_LookupTable — Type

Contains data of a Phase Diagram that is regularly spaced in P & T