Plasticity

Plasticity is a non-linear rheology that is activated once stresses exceed a certain yield criteria.

Implemented laws

The following plastic law are implemented:

GeoParams.MaterialParameters.ConstitutiveRelationships.DruckerPrager — Type

DruckerPrager(ϕ=30, Ψ=0, C=10e6Pa)

Sets parameters for Drucker-Prager plasticity, where the yield stress $\sigma_{y}$ is computed by

\[ \sigma_{y} = (P-P_f)\tan(ϕ) + C\]

with $\phi$ being the friction angle (in degrees), $C$ cohesion, $P$ dynamic pressure and $P_f$ the fluid pressure (both positive under compression).

Yielding occurs when the second invariant of the deviatoric stress tensor, $\tau_{II}=(0.5\tau_{ij}\tau_{ij})^{0.5}$ touches the yield stress. This can be computed with the yield function $F$ and the plastic flow potential $Q$, which are respectively given by

\[ F = \tau_{II} - \cos(ϕ)C - \sin(ϕ)(P-P_f)\]

\[ Q = \tau_{II} - \sin(Ψ)(P-P_f) \]

Here, Ψ is the dilation angle, which must be zero for incompressible setups.

Plasticity is activated when $F(\tau_{II}^{trial})$ (the yield function computed with a trial stress) is >0. In that case, plastic strainrate $\dot{\varepsilon}^{pl}_{ij}$ is computed by:

\[ \dot{\varepsilon}^{pl}_{ij} =\dot{\lambda} {\partial Q \over \partial \sigma_{ij}}\]

where $\dot{\lambda}$ is a (scalar) that is nonzero and chosen such that the resulting stress gives $F(\tau_{II}^{final})=0$, and $\sigma_{ij}=-P + \tau_{ij}$ denotes the total stress tensor.

GeoParams.MaterialParameters.ConstitutiveRelationships.DruckerPrager_regularised — Type

DruckerPrager_regularised(ϕ=30, Ψ=0, C=10e6Pa, η_vp=1e20Pa*s)

Sets parameters for reularised Drucker-Prager plasticity, where the yield stress $\sigma_{y}$ is computed by

\[ \sigma_{y} = (P-P_f)\tan(ϕ) + C + 2η_vpε̇II_pl \]

with $\phi$ being the friction angle (in degrees), $C$ cohesion, $P$ dynamic pressure, $P_f$ the fluid pressure (both positive under compression), $η_vp$ the regularization viscosity and $ε̇II_pl$ the invariant of the plastic strainrate

Yielding occurs when the second invariant of the deviatoric stress tensor, $\tau_{II}=(0.5\tau_{ij}\tau_{ij})^{0.5}$ touches the yield stress. This can be computed with the yield function $F$ and the plastic flow potential $Q$, which are respectively given by

\[ F = \tau_{II} - \cos(ϕ)C - \sin(ϕ)(P-P_f) - 2 \eta_{vp} \dot{\varepsilon}ε̇^{pl}_{II}\]

\[ Q = \tau_{II} - \sin(Ψ)(P-P_f) \]

Here, Ψ is the dilation angle, which must be zero for incompressible setups.

Plasticity is activated when $F(\tau_{II}^{trial})$ (the yield function computed with a trial stress) is >0. In that case, plastic strainrate $\dot{\varepsilon}^{pl}_{ij}$ is computed by:

\[ \dot{\varepsilon}^{pl}_{ij} =\dot{\lambda} {\partial Q \over \partial \sigma_{ij}}\]

where $\dot{\lambda}$ is a (scalar) that is nonzero and chosen such that the resulting stress gives $F(\tau_{II}^{final})=0$, and $\sigma_{ij}=-P + \tau_{ij}$ denotes the total stress tensor.

Computational routines

Usually, plasticity should be defined as part of a CompositeRheology structure and calculations can be done as with all other rheology computations by using compute_τII.