Seismic velocity

Methods

Seismic velocity can specified in a number of ways

GeoParams.MaterialParameters.SeismicVelocity.ConstantSeismicVelocity — Type

ConstantSeismicVelocity(Vp=8.1 km/s, Vs=4.5km/s)

Set a constant seismic P and S-wave velocity:

\[ V_p = cst\]

\[ V_s = cst\]

where $V_p, V_s$ are the P-wave and S-wave velocities [$km/s$].

In addition, you can use phase diagram lookup tables to compute seismic velocities as a function of pressure and temperature.

Computational routines

To evaluate seismic velocity within a user routine, use this:

GeoParams.MaterialParameters.SeismicVelocity.compute_swave_velocity! — Function

compute_swave_velocity!(Vs_array::AbstractArray{<:AbstractFloat}, P::AbstractArray{<:AbstractFloat},T::AbstractArray{<:AbstractFloat}, s::PhaseDiagram_LookupTable)

In-place computation of S-wave velocity as a function of T,P, in case we are using a lookup table.

compute_swave_velocity!(Vs_array::AbstractArray{<:AbstractFloat}, Phases::AbstractArray{<:Integer}, P::AbstractArray{<:AbstractFloat},T::AbstractArray{<:AbstractFloat}, MatParam::AbstractArray{<:AbstractMaterialParamsStruct})

In-place computation of S-wave velocity Vp for the whole domain and all phases, in case a vector with phase properties MatParam is provided, along with P and T arrays. This assumes that the Phase of every point is specified as an Integer in the Phases array.

compute_swave_velocity!(Vs_array::AbstractArray{<:AbstractFloat,N}, PhaseRatios::AbstractArray{<:AbstractFloat, M}, P::AbstractArray{<:AbstractFloat,N},T::AbstractArray{<:AbstractFloat,N}, MatParam::AbstractArray{<:AbstractMaterialParamsStruct})

In-place computation of seismic S-wave velocity Vs for the whole domain and all phases, in case a vector with phase properties MatParam is provided, along with P and T arrays. This assumes that the PhaseRatio of every point is specified as an Integer in the PhaseRatios array, which has one dimension more than the data arrays (and has a phase fraction between 0-1)

GeoParams.MaterialParameters.SeismicVelocity.compute_pwave_velocity! — Function

compute_pwave_velocity!(Vp_array::AbstractArray{<:AbstractFloat}, P::AbstractArray{<:AbstractFloat},T::AbstractArray{<:AbstractFloat}, s::PhaseDiagram_LookupTable)

In-place computation of P-wave velocity as a function of T,P, in case we are using a lookup table.

compute_pwave_velocity!(Vp_array::AbstractArray{<:AbstractFloat}, Phases::AbstractArray{<:Integer}, P::AbstractArray{<:AbstractFloat},T::AbstractArray{<:AbstractFloat}, MatParam::AbstractArray{<:AbstractMaterialParamsStruct})

In-place computation of P-wave velocity Vp for the whole domain and all phases, in case a vector with phase properties MatParam is provided, along with P and T arrays. This assumes that the Phase of every point is specified as an Integer in the Phases array.

compute_pwave_velocity!(Vp_array::AbstractArray{<:AbstractFloat,N}, PhaseRatios::AbstractArray{<:AbstractFloat, M}, P::AbstractArray{<:AbstractFloat,N},T::AbstractArray{<:AbstractFloat,N}, MatParam::AbstractArray{<:AbstractMaterialParamsStruct})

In-place computation of seismic P-wave velocity Vp for the whole domain and all phases, in case a vector with phase properties MatParam is provided, along with P and T arrays. This assumes that the PhaseRatio of every point is specified as an Integer in the PhaseRatios array, which has one dimension more than the data arrays (and has a phase fraction between 0-1)

GeoParams.MaterialParameters.SeismicVelocity.compute_pwave_velocity — Function

compute_pwave_velocity(P,T, s::PhaseDiagram_LookupTable)

Interpolates Vp as a function of T,P

GeoParams.MaterialParameters.SeismicVelocity.compute_swave_velocity — Function

compute_swave_velocity(P,T, s::PhaseDiagram_LookupTable)

Interpolates Vs as a function of T,P

Seismic velocity correction for partial melt

Methods

The routine uses the reduction formulation of Clark et al., (2017) and is based on the equilibrium geometry model for the solid skeleton of Takei et al., 1998.

Computational routines

To compute melt-content based correction for seismic waves velocities, you can use:

GeoParams.MaterialParameters.SeismicVelocity.melt_correction — Function

    Vp_cor,Vs_cor = melt_correction(  Kb_L, Kb_S, Ks_S, ρL, ρS, Vp0, Vs0, ϕ, α)

Corrects P- and S-wave velocities if the rock is partially molten.

Input:

Kb_L: adiabatic bulk modulus of melt

Kb_S: adiabatic bulk modulus of the solid phase

Ks_S: shear modulus of the solid phase

ρL : density of the melt

ρS : density of the solid phase

Vp0 : initial P-wave velocitiy of the solid phase

Vs0 : initial S-wave velocitiy of the solid phase

ϕ : melt volume fraction

α : contiguity coefficient defining the geometry of the solid framework (contiguity) 0.0 (layered melt distributed) < 0.1 (grain boundary melt) < 1.0 (melt in separated bubble pockets)

Output:

Vp_cor,Vs_cor : corrected P-wave and S-wave velocities for melt fraction

The routine uses the reduction formulation of Clark & Lesher, (2017) and is based on the equilibrium geometry model for the solid skeleton of Takei et al., 1998.

References:

Takei (1998) Constitutive mechanical relations of solid-liquid composites in terms of grain-boundary contiguity, Journal of Geophysical Research: Solid Earth, Vol(103)(B8), 18183–18203

Clark & Lesher (2017) Elastic properties of silicate melts: Implications for low velocity zones at the lithosphere-asthenosphere boundary. Science Advances, Vol 3 (12), e1701312

Seismic velocity correction for anelasticity

Methods

The routine uses the reduction formulation of karato (1993), using the quality factor formulation from Behn et al. (2009)

Computational routines

To compute a correction of S-wave velocity for anelasticity, use this:

GeoParams.MaterialParameters.SeismicVelocity.anelastic_correction — Function

    Vs_anel = anelastic_correction(water::Int64, Vs0::Float64,P::Float64,T::Float64)

This routine computes a correction of S-wave velocity for anelasticity

Input:

water: water flag, 0 = dry; 1 = dampened; 2 = water saturated

Vs0 : S-wave velocitiy of the solid phase (with or without melt correction)

P : pressure given in kbar

T : temperature given in °C

Output:

Vs_anel : corrected S-wave velocity for anelasticity

The routine uses the reduction formulation of Karato (1993), using the quality factor formulation from Behn et al. (2009)

References:

Karato, S. I. (1993). Importance of anelasticity in the interpretation of seismic tomography. Geophysical research letters, 20(15), 1623-1626.

Behn, M. D., Hirth, G., & Elsenbeck II, J. R. (2009). Implications of grain size evolution on the seismic structure of the oceanic upper mantle. Earth and Planetary Science Letters, 282(1-4), 178-189.