Seismic velocity

Methods

Seismic velocity can specified in a number of ways

GeoParams.MaterialParameters.SeismicVelocity.ConstantSeismicVelocity — Type

ConstantSeismicVelocity(Vp=8.1 km/s, Vs=4.5km/s)

Set a constant seismic P and S-wave velocity:

\[ V_p = cst \ V_s = cst\]

where $V_p, V_s$ are the P-wave and S-wave velocities [$km/s$].

In addition, you can use phase diagram lookup tables to compute seismic velocities as a function of pressure and temperature.

Computational routines

To evaluate seismic velocity within a user routine, use this:

GeoParams.MaterialParameters.SeismicVelocity.compute_swave_velocity! — Function

compute_swave_velocity!(Vs_array::AbstractArray{<:AbstractFloat}, P::AbstractArray{<:AbstractFloat},T::AbstractArray{<:AbstractFloat}, s::PhaseDiagram_LookupTable)

In-place computation of S-wave velocity as a function of T,P, in case we are using a lookup table.

compute_swave_velocity!(Vs_array::AbstractArray{<:AbstractFloat}, Phases::AbstractArray{<:Integer}, P::AbstractArray{<:AbstractFloat},T::AbstractArray{<:AbstractFloat}, MatParam::AbstractArray{<:AbstractMaterialParamsStruct})

In-place computation of S-wave velocity Vp for the whole domain and all phases, in case a vector with phase properties MatParam is provided, along with P and T arrays. This assumes that the Phase of every point is specified as an Integer in the Phases array.

compute_swave_velocity!(Vs_array::AbstractArray{<:AbstractFloat,N}, PhaseRatios::AbstractArray{<:AbstractFloat, M}, P::AbstractArray{<:AbstractFloat,N},T::AbstractArray{<:AbstractFloat,N}, MatParam::AbstractArray{<:AbstractMaterialParamsStruct})

In-place computation of seismic S-wave velocity Vs for the whole domain and all phases, in case a vector with phase properties MatParam is provided, along with P and T arrays. This assumes that the PhaseRatio of every point is specified as an Integer in the PhaseRatios array, which has one dimension more than the data arrays (and has a phase fraction between 0-1)

GeoParams.MaterialParameters.SeismicVelocity.compute_pwave_velocity! — Function

compute_pwave_velocity!(Vp_array::AbstractArray{<:AbstractFloat}, P::AbstractArray{<:AbstractFloat},T::AbstractArray{<:AbstractFloat}, s::PhaseDiagram_LookupTable)

In-place computation of P-wave velocity as a function of T,P, in case we are using a lookup table.

compute_pwave_velocity!(Vp_array::AbstractArray{<:AbstractFloat}, Phases::AbstractArray{<:Integer}, P::AbstractArray{<:AbstractFloat},T::AbstractArray{<:AbstractFloat}, MatParam::AbstractArray{<:AbstractMaterialParamsStruct})

In-place computation of P-wave velocity Vp for the whole domain and all phases, in case a vector with phase properties MatParam is provided, along with P and T arrays. This assumes that the Phase of every point is specified as an Integer in the Phases array.

compute_pwave_velocity!(Vp_array::AbstractArray{<:AbstractFloat,N}, PhaseRatios::AbstractArray{<:AbstractFloat, M}, P::AbstractArray{<:AbstractFloat,N},T::AbstractArray{<:AbstractFloat,N}, MatParam::AbstractArray{<:AbstractMaterialParamsStruct})

In-place computation of seismic P-wave velocity Vp for the whole domain and all phases, in case a vector with phase properties MatParam is provided, along with P and T arrays. This assumes that the PhaseRatio of every point is specified as an Integer in the PhaseRatios array, which has one dimension more than the data arrays (and has a phase fraction between 0-1)

GeoParams.MaterialParameters.SeismicVelocity.compute_pwave_velocity — Function

compute_pwave_velocity(P,T, s::PhaseDiagram_LookupTable)

Interpolates Vp as a function of T,P

GeoParams.MaterialParameters.SeismicVelocity.compute_swave_velocity — Function

compute_swave_velocity(P,T, s::PhaseDiagram_LookupTable)

Interpolates Vs as a function of T,P