Shear heating

Heat caused by non-recoverable deformation can be specified in

GeoParams.MaterialParameters.Shearheating.ConstantShearheating — Type

ConstantShearheating(Χ=0.0NoUnits)

Set the shear heating efficiency [0-1] parameter

\[Χ = cst\]

where $\Chi$ is the shear heating efficiency [NoUnits]

Shear heating is computed as

\[H_s = \Chi \cdot \tau_{ij}(\dot{\varepsilon}_{ij} - \dot{\varepsilon}^{el}_{ij})\]

To compute, use this:

GeoParams.MaterialParameters.Shearheating.ComputeShearheating — Function

H_s = ComputeShearheating(τ, ε, ε_el,  s:<AbstractShearheating)

Computes the shear heating source term

\[H_s = \Chi \cdot \tau_{ij} ( \dot{\varepsilon}_{ij} - \dot{\varepsilon}^{el}_{ij})\]

Parameters

$\Chi$ : The efficiency of shear heating (between 0-1)

$\tau_{ij}$ : The full deviatoric stress tensor [4 components in 2D; 9 in 3D]

$\dot{\varepsilon}_{ij}$ : The full deviatoric strainrate tensor

$\dot{\varepsilon}^{el}_{ij}$ : The full elastic deviatoric strainrate tensor

H_s = ComputeShearheating(τ, ε,  s:<AbstractShearheating)

Computes the shear heating source term when there is no elasticity

\[H_s = \Chi \cdot \tau_{ij} \dot{\varepsilon}_{ij} \]

Parameters

$\Chi$ : The efficiency of shear heating (between 0-1)

$\tau_{ij}$ : The full deviatoric stress tensor [4 components in 2D; 9 in 3D]

$\dot{\varepsilon}_{ij}$ : The full deviatoric strainrate tensor

GeoParams.MaterialParameters.Shearheating.ComputeShearheating! — Function

ComputeShearheating!(H_s, τ, ε, ε_el,  s:<AbstractShearheating)

Computes the shear heating source term in-place

\[H_s = \Chi \cdot \tau_{ij} ( \dot{\varepsilon}_{ij} - \dot{\varepsilon}^{el}_{ij})\]

Parameters

$\Chi$ : The efficiency of shear heating (between 0-1)

$\tau_{ij}$ : The full deviatoric stress tensor [4 components in 2D; 9 in 3D]

$\dot{\varepsilon}_{ij}$ : The full deviatoric strainrate tensor

$\dot{\varepsilon}^{el}_{ij}$ : The full elastic deviatoric strainrate tensor

NOTE: The shear heating terms require the full deviatoric stress & strain rate tensors, i.e.:

\[2D: \tau_{ij} = \left( \begin{matrix} \tau_{xx} & \tau_{xz} \\ \tau_{zx} & \tau_{zz} \end{matrix} \right)\]

Since $\tau_{zx}=\tau_{xz}$, most geodynamic codes only take one of the terms into account; shear heating requires all components to be used!

ComputeShearheating!(H_s, τ, ε,  s:<AbstractShearheating)

Computes the shear heating source term H_s in-place when there is no elasticity

\[H_s = \Chi \cdot \tau_{ij} \dot{\varepsilon}_{ij} \]

Parameters

$\Chi$ : The efficiency of shear heating (between 0-1)

$\tau_{ij}$ : The full deviatoric stress tensor [4 components in 2D; 9 in 3D]

$\dot{\varepsilon}_{ij}$ : The full deviatoric strainrate tensor